requirements.txt
torch>=2.1.0
torchvision>=0.16.0
pytorch-lightning>=2.1.0
transformers>=4.36.0
albumentations>=1.3.1
opencv-python>=4.8.0
timm>=0.9.12
accelerate>=0.25.0
onnx>=1.15.0
insightface>=0.7.3
facenet-pytorch>=2.5.3
deepface>=0.0.79
h2o>=3.44.0
tensorboard>=2.15.0
wandb>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
scikit-learn>=1.3.0
